Get the parser with default values
This is, because some languages, like c# do not work with the filetype
for some reason. Trying to get the parser of "cs" crashes the plugin.

The default values are the current buffer (this is always the case) and
the current languages (this is also always the case). So nothing that
might that should make any problems on first sight.