Use named node instead of random ones by index
After having some problems in different languages and not finding an
easy way to get a specific named node, I added a function to fetch the
wanted named node myself.

This allows the plugin to be more stable and less likely to get a node
that is not the correct one.

Example: Normally the 2nd node is the name of a class, but some
languages have a prefix (async, static, etc.) in front of the class
name, meaning that now the correct one would be the 3rd or 4th node,
which I do not want, because treesitter should handle that not me.