Use line content as a fallback of function name
Not all languages have named nodes, but still have the same structure.
Instead of not working in them (vimscript, lua, etc) we just fallback
into the line content showing what exists on the line.

This should be good enough for now. The approach of having sane
fallbacks should be a good one and I do not see any problems with this
approach so far.