Handle all kinds of structure recursively
Because structures like variables (arrow functions) etc. might have in
itself functions that are already handled by the recursive function, it
is easier to just handle them by calling the recursive function instead
of just rewriting them again.

After the refactor most parts were able to be grouped and these groups
were refactored again to allow for a (hopefully) cleaner code.

One problem is the case of edgecases that might occur in different
languages differently (function definition). In lua it is a variable
that is assigned and needs to know its parent, in php its a normal
function meaning that no parent is needed.
To solve this, both cases are used everytime resulting in redundant
matches. This is not pretty but better than not showing anything in some
cases at all (in my opinion).

Later on a sort and filter function can make this result less obvious.