Sort and filter resulting tables
It shouldnt make a problem to sort and to filter duplicates. Most of the
time this will do nothing.

It is mainly needed for php/lua (and probably others) that have nodes
that are used in multiple parts and because of it adding it multiple
times to the output. This will not fix it, but make it useable (sorting
everything even if its duplicated is fine & removing 1:1 copies also
helps a bit).
This does not help against tables that have the same data logically but
the information is different (lua, where one is just "function" the
other one is the whole line). Both are on the same line, but I do not
want to remove either of them for now.
Might be that I find a good fix for it in the future.

This bug can be seen in one or a few commits before this one.