Remove filter of resulting table
This was mainly done as a quickfix of an error (see previous commits).

This will be somewhat fixed in the next, by checking if the parent has
the info that is needed, if not it will be ignored.
The other side of the problem (lua mainly) will be ignored and is still
there, but should not be as problematic as the other one.