Check and use existing fuzzy finder
Firstly fzf uses the lua formatted string (it is the same), because then
lua plugins can use it natively and do not need to call a vim function
which calls a lua function, or need to reimplement the same logic
multiple times.

Additionally switch between the different fuzzy finder and display
in case there is non installed to the user.