Refactor internal structure of plugin
TLDR
* internal naming changed from treesitter-current-function to tscf
  * because autoload and easier to reason with
  * lua structure also changed to keep same naming at all places
* removed selector part from plugin folder (into its own subdir)
  * added autoload folder for vim selectors

Remove the selector implementation of fzf out of the plugin entrypoint
(plugin folder). Put it into the autoload folder.
Because of that we want to use the autoloaded functions instead with
filepath#subdir#functionname() instead, these seem to not work with
dashes (of treesitter-current-functions).
Because the name was already way to long for programming purpose it was
shortened to "tscf".
This means the vim side of things uses tscf instead of the long name,
this should be kept over the whole plugin, meaning the lua needs to be
udpated as well, resulting in a bigger change than expected.