Rename entry point function
Mainly because the naming was the same as the function that generates
the output, which is consumed by the selector.

This makes it more difficult to see which one is mentioned in
documentation etc.

Additionally the name "get" does not really make sense in that content,
the treesitter function gets the content, this functions allows
selection.