Update force selector command to take an argument
Instead of having multiple commands that just have the selector in the
name, I added a command that can take an argument and this will be
passed to the internal function instead.
The command forces one argument and it also adds tab completion for
easier access.

This will make later integration of other selectors easier and I think
this is overall a better way of handling it instead of adding tens of
redundant commands.

Also updated the doc accordingly.