Handle case of no ts parser silently
Instead of crashing the call and throwing an error to the user, just
send a message to the user and return nothing, resulting the plugin to
still run normally and showing the case, that nothing exists.

Now the case of no parser exists is not an error anymore but a normal
possible state that the call can be in.